Revert "feat(test runner): test.reset() to reset options to default/config value (#18561)"
This reverts commit 6fef227f43f122d45327768addc19bb96db0f6d5.
Not shipping in 1.28.
