+++ packages/playwright-test/src/types.ts
@@ -24,7 +24,6 @@ export type { Location } from '../types/testReporter';
 export type FixturesWithLocation = {
   fixtures: Fixtures;
   location: Location;
-  fromConfig?: boolean;
 };
 export type Annotation = { type: string, description?: string };
