+++ tests/playwright-test/test-extend.spec.ts
@@ -225,3 +225,54 @@ test('test._extendTest should print nice message when used as extend', async ({ 
   expect(result.passed).toBe(0);
   expect(result.output).toContain('Did you mean to call test.extend() with fixtures instead?');
 });
+
+test('test.use() with undefined should not be ignored', async ({ runInlineTest }) => {
+  const result = await runInlineTest({
+    'playwright.config.ts': `
+      module.exports = {
+        use: { option1: 'config' },
+      };
+    `,
+    'a.test.js': `
+      const test = pwt.test.extend({
+        option1: [ 'default', { option: true } ],
+        option2: [ 'default', { option: true } ],
+      });
+      test('test1', async ({ option1, option2 }) => {
+        console.log('test1: option1=' + option1);
+        console.log('test1: option2=' + option2);
+      });
+
+      test.describe('', () => {
+        test.use({ option1: 'foo', option2: 'foo' });
+        test('test2', async ({ option1, option2 }) => {
+          console.log('test2: option1=' + option1);
+          console.log('test2: option2=' + option2);
+        });
+
+        test.describe('', () => {
+          test.use({ option1: undefined, option2: undefined });
+          test('test3', async ({ option1, option2 }) => {
+            console.log('test3: option1=' + option1);
+            console.log('test3: option2=' + option2);
+          });
+        });
+      });
+
+      test.extend({ option1: undefined, option2: undefined })('test4', async ({ option1, option2 }) => {
+        console.log('test4: option1=' + option1);
+        console.log('test4: option2=' + option2);
+      });
+    `,
+  });
+  expect(result.exitCode).toBe(0);
+  expect(result.passed).toBe(4);
+  expect(result.output).toContain('test1: option1=config');
+  expect(result.output).toContain('test1: option2=default');
+  expect(result.output).toContain('test2: option1=foo');
+  expect(result.output).toContain('test2: option2=foo');
+  expect(result.output).toContain('test3: option1=config');
+  expect(result.output).toContain('test3: option2=default');
+  expect(result.output).toContain('test4: option1=config');
+  expect(result.output).toContain('test4: option2=default');
+});
